spacecmd: Skip time prompt when running with --yes
I think it's a fair assumption that people running spacecmd with --yes
expect the questions to be skipped. Therefore for all the recently added
start_time options, that is in:
  do_system_reboot()
  do_system_installpackage()
  do_system_removepackage()
  do_system_upgradepackage()
  do_system_deployconfigfiles()
  do_system_syncpackages()
  do_system_schedulehardwarerefresh()
  do_system_schedulepackagerefresh()
skip the "Start Time [now]:" question if running with --yes, and use the
default "now" time specifier.

This has a benefit of restoring backwards compatibility for people who
run this in batch jobs like this:

  $ for h in $HOSTS; do spacecmd -y -- system_reboot $h; done

and expect no additional prompts.

Signed-off-by: Jakub Jankowski <shasta@toxcorp.com>