1342977 - Prevent ORA-01878 on repository sync.
This prevents ORA-01878 on Oracle, and incorrect timezone insertion on
PostgreSQL when repo sync is triggered from the web UI. It also prevents
timestamps being off by one hour when spacewalk-repo-sync is run from
the cmdline.

See Bugzilla for details.

This is a symptomatic patch. The Python ORM should really be supplying
the timezone with submitted timestamps in order to prevent future
timestamps being submitted in UTC causing problems.