spacecmd: user: allow more than one group at a time
The backend can take a list of groups for the following four operations:
 addAssignedSystemGroups
 addDefaultSystemGroups
 removeAssignedSystemGroups
 removeDefaultSystemGroups
and spacecmd help text suggests this is also possible with
user_{add,remove}{default,}group - but because the code checks for
exactly two arguments, it fails when given more than two.

Fix it by relaxing the check to "less than two args".

Signed-off-by: Jakub Jankowski <shasta@toxcorp.com>