1447296 - add package_import_skip_changelog option to speed reposync up
Considerable time is taken by spacewalk-repo-sync and inter-server sync
due to database insertions in tables with a high number of rows.

Two of these (rhnPackageChangeLogRec and rhnPackageChangeLogData) only contain
informational data that can be skipped in order to get faster repo-sync times,
at the expense of not finding this data in the Web UI.

This patch adds an rhn.conf flag (package_import_skip_changelog = 1) that will
result in spacewalk-repo-sync and inter-server sync skipping those tables with
noticeable performance gains.

This is particularly useful in tests and demos, and is not recommended in
production.