bz1497216 - [RFE] Add softwarechannel_setsyncschedule --latest
Add support for the `--latest` option on the
`softwarechannel_setsyncshedule` command so that tasko sync tasks
only download the most recent versions rather than the full repository
content.

This brings the spacecmd command closer into line with the features
available via the webinterface and allows use of the latest option when
setting sync schedules programmatically.

Fixes #1497216