Update Python 2 dependency declarations to new packaging standards
This package uses names with ambiguous python- prefix in requirements.

According to Fedora Packaging guidelines for Python, packages must use names with either python2- or python3- prefix in requirements where available.
We are aiming to rename python-* dependencies to python2-*, so we can later switch the python-* namespace to Python 3.

This PR is part of Fedora's Switch to Python 3 effort.

Note that, although this PR was created automatically, we will respond to any comments or issues which you might find with it. We will keep the PR open for review for a week, and if there's no feedback we'll merge it. There is no need to build the package right after merge, this change can wait for the next package rebuild.
The local mock build passed, for Koji scratch build please see simple-koji-ci result.

Taken from:
https://src.fedoraproject.org/rpms/python-gzipstream/pull-request/1#request_diff