Fix crash in satellite_tools on RH/Fedora/Mageia urlgrabber
The 'authtoken' parameter isn't available in non-SUSE 'urlgrabber', which
causes it to crash when it attempts to execute this code. Checking for
the existence of the key in the 'params' dictionary prevents the failure
while preserving compatibility with SUSE systems.

(cherry picked from commit ba3c90b309c8da4b5142c5b4650b2d2128ed5840)