client/rhel: Enable DNF plugin for Mageia 6+ and openSUSE Leap 15.0+
Since Mageia 6 and openSUSE Leap 15.0, DNF has been available in the
distributions for usage. In addition, Mageia 6 offers DNF as a fully
supported package manager option, with plans to transition to it
as the only system package manager in the next release.

This will permit Mageia and openSUSE machines using DNF to work with
Spacewalk.

In addition, the yum plugin was trivially adapted to be able to build
for Mageia, too, since Mageia has a working yum now, too.